% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{count_fire_locations}
\alias{count_fire_locations}
\title{Detect whether a fire reached given locations}
\usage{
count_fire_locations(outdir, prefix, pts)
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}

\item{pts}{A matrix or data frame of point coordinates (X-Y order). The units
and coordinate reference sytem are assumed to match the FARSITE output rasters.}
}
\value{
The number of locations reached by the fire.
}
\description{
This is a convenience function to test if a fire reached one or more target locations
represented by a set of point coordinates. Each point is tested by checking the
FARSITE fire arrival time raster at that location.
}
