% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{create_ignition_polygon}
\alias{create_ignition_polygon}
\title{Create an ignition point shapefile}
\usage{
create_ignition_polygon(path, xy, width, crs = sf::NA_crs_)
}
\arguments{
\item{path}{Path for the output shapefile.}

\item{xy}{Two element vector with X and Y ordinates of ignition location.}

\item{width}{Width of the polygon to create.}

\item{crs}{Optional coordinate reference sytem (e.g. EPSG integer code).}
}
\value{
A spatial data frame (\code{sf} object) with the ignition polygon.
}
\description{
FARSITE requires a polygon shapefile for ignition location. This function
takes the coordinates for a point, creates a square polygon around it
of specified width, and saves this as a shapefile.
}
