% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{create_input_file}
\alias{create_input_file}
\title{Create a FARSITE input file}
\usage{
create_input_file(start.time, end.time, path.fms, raws.elev, raws.units,
  path.raws, path.fmd = NULL, path.ros.adj = NULL)
}
\arguments{
\item{start.time}{Character string for start date and time, e.g. '08 28 0000'.}

\item{end.time}{Character string for end date and time, e.g. '08 29 2359'.}

\item{path.fms}{Path to a file of tabular data for fuel moisture in the
six-column format expected by FARSITE. Note, only the data records should
be included in the file without any header information.}

\item{raws.elev}{Elevation value.}

\item{raws.units}{Units of measurement: one of 'Metric' or 'English'.}

\item{path.raws}{Path to a file of tabular RAWS weather data in the
ten-column format expected by FARSITE. Note, only the data records should
be included in the file without any header information.}

\item{path.fmd}{Optional path to a custom fuels file.}

\item{path.ros.adj}{Optional path to the ROS adjustment file.}
}
\value{
Contents of the input file as a single character string.
}
\description{
This function creates the contents for a FARSITE input file from a template
(package file \code{extdata/input_template.txt}) and a set of values for
variable components.
}
