% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{format_run_args}
\alias{format_run_args}
\title{Format FARSITE run command arguments}
\usage{
format_run_args(path.lcp, path.input, path.ignition, path.outdir, prefix)
}
\arguments{
\item{path.lcp}{Path to the landscape (.lcp) file.}

\item{path.input}{Path to the FARSITE input file.}

\item{path.ignition}{Path to the shapefile of ignition locations.}

\item{out.dir}{Directory for FARSITE output files.}

\item{out.prefix}{File name prefix for FARSITE output files.}
}
\value{
FARSITE run command arguments as a single character string.
}
\description{
Format FARSITE run command arguments
}
