% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_crownfire_area}
\alias{get_crownfire_area}
\title{Get the area burnt by surface, passive-crown or active-crown fire}
\usage{
get_crownfire_area(outdir, prefix)
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}
}
\value{
A named vector of areas ('surface', 'passive', 'active') in
  user-defined map units (e.g. square metres).
}
\description{
This function determines the area burnt by each of three categories of fire:
surface fire; passive crown fire; and active crown fire. The areas are
calculated from crown fire raster generated by FARSITE. The precision of the
area estimate will depend on the cell resolution specified in the FARSITE
input file.
}
