% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_final_time}
\alias{get_final_time}
\title{Get last fire time step}
\usage{
get_final_time(outdir, prefix)
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}
}
\value{
A named numeric vector with elements 'month', 'day' and 'hour'.
}
\description{
Retrieves the month, day and time (24 hour format) of the last fire
perimeter update. This can be used, for example, to check if the fire was still
active at the end of the burn period specified of the simulation.
}
