% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_fire_area}
\alias{get_fire_area}
\title{Get final fire area}
\usage{
get_fire_area(outdir, prefix)
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}
}
\value{
Fire area in user-defined map units (e.g. square metres).
}
\description{
This function determines the final fire area from the raster of fire arrival
times generated by FARSITE. The precision of the area estimate will depend on
the cell resolution specified in the FARSITE input file.
}
