% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_fire_summary}
\alias{get_fire_summary}
\title{Get summary statistics for a simulated fire}
\usage{
get_fire_summary(outdir, prefix)
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}
}
\description{
Get summary statistics for a simulated fire
}
