% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_intensity_hist}
\alias{get_intensity_hist}
\title{Summarize fire intensity}
\usage{
get_intensity_hist(outdir, prefix, breaks = c(0, 350, 1700, 3500, Inf))
}
\arguments{
\item{outdir}{Full path to the directory containing output files.}

\item{prefix}{Prefix used for the output file names.}

\item{breaks}{A vector of breaks to define intensity classes. Classes are
treated as left-open / right-closed.}
}
\value{
A named list with elements: 'breaks' and 'area'. Area values are
  expressed in user-defined map units (e.g. square metres).
}
\description{
This function summarizes fire intensity by calculating the area of each of a
set of intensity classes, based on the raster of fire intensity generated by
FARSITE. The precision of the area estimate will depend on the cell
resolution specified in the FARSITE input file.
}
