% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_farsite}
\alias{run_farsite}
\title{Run a FARSITE simulation}
\usage{
run_farsite(lcp, input, ignition, outdir, prefix = "sim",
  cmd = "farsite")
}
\arguments{
\item{lcp}{Full path to the landscape file.}

\item{input}{Full path to the input file.}

\item{ignition}{Full path to the ignition shapefile.}

\item{outdir}{Full path to the directory for output files.}

\item{prefix}{Prefix to use for output file names.}

\item{cmd}{System command to run FARSITE.}
}
\value{
An integer indicating success (zero) or failure (non-zero). 127
  indicates that FARSITE could not be run for some reason. 124 indicates
  that the simulation timed out.
}
\description{
Run a FARSITE simulation
}
